backend/requirements.txt
# Backend dependencies
flask==2.3.3
flask-cors==4.0.0
tensorflow==2.16.1
numpy==1.26.4
Pillow==10.2.0
gunicorn==21.2.0
python-dotenv==1.0.1
requests==2.31.0
scikit-learn==1.4.1
pandas==2.2.1
matplotlib==3.8.3
scipy==1.13.0
opencv-python==4.9.0.80
json5==0.9.17
tqdm==4.66.2

# Extra utilities for deployment
waitress==2.1.2
pytest==7.4.4
black==24.2.0
isort==5.13.2
flake8==7.0.0